refactor: Reorganize application component file structure
Moves core OS-like applications (AppStore, Settings, FileExplorer) into the window/components directory.

Moves system utility applications (Notebook, SFTP, Terminus) into a new window/components/apps directory.

Updates all import paths across the application to reflect the new file locations. This extensive refactoring follows a user request to change the project's folder hierarchy.
