+++ components/apps/index.ts
@@ -3,19 +3,19 @@
 import type { AppDefinition } from '../../window/types';
 
 import { appDefinition as aboutAppDefinition } from './AboutApp';
-import { appDefinition as fileExplorerAppDefinition } from './FileExplorerApp';
+import { appDefinition as fileExplorerAppDefinition } from '../../window/components/FileExplorerApp';
 import { appDefinition as geminiChatAppDefinition } from './GeminiChatApp';
 import { appDefinition as hyperAppDefinition } from './HyperApp';
-import { appDefinition as notebookAppDefinition } from './NotebookApp';
-import { appDefinition as settingsAppDefinition } from './SettingsApp';
+import { appDefinition as notebookAppDefinition } from '../../window/components/apps/NotebookApp';
+import { appDefinition as settingsAppDefinition } from '../../window/components/SettingsApp';
 import { appDefinition as chromeAppDefinition } from './ChromeApp';
 import { appDefinition as chrome2AppDefinition } from './Chrome2App';
 import { appDefinition as chrome3AppDefinition } from './Chrome3App'; // Restored Chrome 3
 import { appDefinition as chrome4AppDefinition } from './Chrome4App';
-import { appDefinition as terminusAppDefinition } from './TerminusApp';
+import { appDefinition as terminusAppDefinition } from '../../window/components/apps/TerminusApp';
 import { appDefinition as terminusSshAppDefinition } from './TerminusSshApp';
-import { appDefinition as sftpAppDefinition } from './SFTPApp';
-import { appDefinition as appStoreAppDefinition } from './AppStoreApp';
+import { appDefinition as sftpAppDefinition } from '../../window/components/apps/SFTPApp';
+import { appDefinition as appStoreAppDefinition } from '../../window/components/AppStoreApp';
 import { appDefinition as themeAppDefinition } from './ThemeApp';
 import { appDefinition as propertiesAppDefinition } from './PropertiesApp';
