+++ window/components/AppStore/index.tsx
