+++ window/components/AppStoreApp.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect, useCallback } from 'react';
 import { AppDefinition, AppComponentProps } from '../../window/types';
 import { AppStoreIcon, RefreshIcon } from '../../window/constants';
 import * as FsService from '../../services/filesystemService';
-import { APP_DEFINITIONS } from '.';
+import { APP_DEFINITIONS } from '../../components/apps';
 
 const AppStoreApp: React.FC<AppComponentProps> = ({ setTitle, initialData }) => {
     const [installedAppIds, setInstalledAppIds] = useState<Set<string>>(new Set());
