+++ window/components/Settings/index.tsx
