+++ window/components/apps/Terminus/index.tsx
